+++ app/project-listing/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Added project text-case
+- Added 2 projects: random-id, text-case
 - Added home page to show projects
 - Added app description
 - Resolved url by environment